+++ pygal/svg.py
@@ -68,10 +68,11 @@ class Svg(object):
 
         self.root = etree.Element('svg', **attrs)
         self.root.attrib['id'] = self.id.lstrip('#').rstrip()
-        self.root.attrib['class'] = 'pygal-chart'
+        if graph.classes:
+            self.root.attrib['class'] = ' '.join(graph.classes)
         self.root.append(
             etree.Comment(u(
-                'Generated with pygal %s (%s) ©Kozea 2011-2015 on %s' % (
+                'Generated with pygal %s (%s) ©Kozea 2011-2016 on %s' % (
                     __version__,
                     'lxml' if etree.lxml else 'etree',
                     date.today().isoformat()))))