+++ docs/changelog.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+2.2.2
+=====
+
+* Add classes option.
+* Handle ellipsis in list type configs to auto-extend parent. (Viva python3)
+
+
 2.2.0
 =====